+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentMethods/STPPaymentMethodEnums.swift
@@ -102,6 +102,10 @@ import Foundation
     case mbWay
     /// A Bizum payment method
     case bizum
+    /// A Naver Pay payment method
+    case naverPay
+    /// A PAYCO payment method
+    case payco
     /// An unknown type.
     case unknown
 
@@ -202,6 +206,10 @@ import Foundation
             return "MB WAY"
         case .bizum:
             return "Bizum"
+        case .naverPay:
+            return "Naver Pay"
+        case .payco:
+            return "PAYCO"
         case .cardPresent,
             .unknown:
             return STPLocalizedString("Unknown", "Default missing source type label")
@@ -305,6 +313,10 @@ import Foundation
             return "mb_way"
         case .bizum:
             return "bizum"
+        case .naverPay:
+            return "naver_pay"
+        case .payco:
+            return "payco"
         }
     }