+++ CHANGELOG.md
@@ -3,6 +3,10 @@ PATCH
 
 ## X.Y.Z - changes pending release
 
+### Payments
+* [Added] Added API bindings for Naver Pay.
+* [Added] Added API bindings for PAYCO.
+
 ## 26.8.0 2026-08-24
 ### Payments
 * [Added] Added support for the following FPX banks: Agrobank, Bank of China, and MBSB Bank.